+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_enclave_functions.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"github.com/gammazero/workerpool"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_resource_collectors"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_key_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_value_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
@@ -39,6 +40,20 @@ const (
 	enclaveDumpJsonSerializationIndent = "  "
 	enclaveDumpJsonSerializationPrefix = ""
 )
+
+// Any of these values being nil indicates that the resource doesn't exist
+type enclaveKubernetesResources struct {
+	namespace *apiv1.Namespace
+
+	// Pods are technically not resources that define an enclave, but we need them both
+	//to StopEnclave and to return an EnclaveStatus
+	pods []*apiv1.Pod
+}
+
+
+// ====================================================================================================
+//                                     		Enclave CRUD Methods
+// ====================================================================================================
 
 func (backend *KubernetesKurtosisBackend) CreateEnclave(
 	ctx context.Context,
@@ -79,7 +94,7 @@ func (backend *KubernetesKurtosisBackend) CreateEnclave(
 	enclaveNamespaceName := enclaveNamespaceAttrs.GetName().GetString()
 	enclaveNamespaceLabels := getStringMapFromLabelMap(enclaveNamespaceAttrs.GetLabels())
 
-	_, err = backend.kubernetesManager.CreateNamespace(ctx, enclaveNamespaceName, enclaveNamespaceLabels)
+	enclaveNamespace, err := backend.kubernetesManager.CreateNamespace(ctx, enclaveNamespaceName, enclaveNamespaceLabels)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Failed to create namespace with name '%v' for enclave '%v'", enclaveNamespaceName, enclaveId)
 	}
@@ -98,19 +113,12 @@ func (backend *KubernetesKurtosisBackend) CreateEnclave(
 		return nil, stacktrace.Propagate(err, "An error occurred while trying to get the enclave data volume attributes for the enclave with ID '%v'", enclaveId)
 	}
 
-	persistentVolumeClaimName := enclaveDataVolumeAttrs.GetName()
-	persistentVolumeClaimLabels := enclaveDataVolumeAttrs.GetLabels()
+	persistentVolumeClaimName := enclaveDataVolumeAttrs.GetName().GetString()
+	persistentVolumeClaimLabels := getStringMapFromLabelMap(enclaveDataVolumeAttrs.GetLabels())
 
-	enclaveVolumeLabelMap := map[string]string{}
-	for kubernetesLabelKey, kubernetesLabelValue := range persistentVolumeClaimLabels {
-		pvcLabelKey := kubernetesLabelKey.GetString()
-		pvcLabelValue := kubernetesLabelValue.GetString()
-		enclaveVolumeLabelMap[pvcLabelKey] = pvcLabelValue
-	}
-
-	foundVolumes, err := backend.kubernetesManager.GetPersistentVolumeClaimsByLabels(ctx, enclaveNamespaceName, enclaveVolumeLabelMap)
+	foundVolumes, err := backend.kubernetesManager.GetPersistentVolumeClaimsByLabels(ctx, enclaveNamespaceName, persistentVolumeClaimLabels)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting enclave data volumes matching labels '%+v'", enclaveVolumeLabelMap)
+		return nil, stacktrace.Propagate(err, "An error occurred getting enclave data volumes matching labels '%+v'", persistentVolumeClaimLabels)
 	}
 	if len(foundVolumes.Items) > 0 {
 		return nil, stacktrace.NewError("Cannot create enclave with ID '%v' because one or more enclave data volumes for that enclave already exists", enclaveId)
@@ -118,37 +126,50 @@ func (backend *KubernetesKurtosisBackend) CreateEnclave(
 
 	_, err = backend.kubernetesManager.CreatePersistentVolumeClaim(ctx,
 		enclaveNamespaceName,
-		persistentVolumeClaimName.GetString(),
-		enclaveVolumeLabelMap,
+		persistentVolumeClaimName,
+		persistentVolumeClaimLabels,
 		backend.volumeSizePerEnclaveInMegabytes,
 		backend.volumeStorageClassName)
 	if err != nil {
 		return nil, stacktrace.Propagate(err,
 			"Failed to create persistent volume claim in enclave '%v' with name '%v' and storage class name '%v'",
 			enclaveNamespaceName,
-			persistentVolumeClaimName.GetString(),
+			persistentVolumeClaimName,
 			backend.volumeStorageClassName)
 	}
 	shouldDeleteVolume := true
 	defer func() {
 		if shouldDeleteVolume {
-			if err := backend.kubernetesManager.RemovePersistentVolumeClaim(teardownContext, enclaveNamespaceName,persistentVolumeClaimName.GetString()); err != nil {
+			if err := backend.kubernetesManager.RemovePersistentVolumeClaim(teardownContext, enclaveNamespaceName, persistentVolumeClaimName); err != nil {
 				logrus.Errorf(
 					"Creating the enclave didn't complete successfully, so we tried to delete enclave persistent volume claim '%v' " +
 						"that we created but an error was thrown:\n%v",
-					persistentVolumeClaimName.GetString(),
+					persistentVolumeClaimName,
 					err,
 				)
-				logrus.Errorf("ACTION REQUIRED: You'll need to manually remove persistent volume claim with name '%v'!!!!!!!", persistentVolumeClaimName.GetString())
+				logrus.Errorf("ACTION REQUIRED: You'll need to manually remove persistent volume claim with name '%v'!!!!!!!", persistentVolumeClaimName)
 			}
 		}
 	}()
 
-	enclaveObj := enclave.NewEnclave(enclaveId, enclave.EnclaveStatus_Empty)
+	enclaveResources := &enclaveKubernetesResources{
+		namespace: enclaveNamespace,
+		pods: []*apiv1.Pod{},
+	}
+	enclaveObjsById, err := getEnclaveObjectsFromKubernetesResources(map[enclave.EnclaveID]*enclaveKubernetesResources{
+		enclaveId: enclaveResources,
+	})
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred converting the new enclave's Kubernetes resources to enclave objects")
+	}
+	resultEnclave, found := enclaveObjsById[enclaveId]
+	if !found {
+		return nil, stacktrace.NewError("Successfully converted the new enclave's Kubernetes resources to an enclave object, but the resulting map didn't have an entry for enclave ID '%v'", enclaveId)
+	}
 
 	shouldDeleteVolume = false
 	shouldDeleteNamespace = false
-	return enclaveObj, nil
+	return resultEnclave, nil
 }
 
 func (backend *KubernetesKurtosisBackend) GetEnclaves(
@@ -158,30 +179,121 @@ func (backend *KubernetesKurtosisBackend) GetEnclaves(
 	map[enclave.EnclaveID]*enclave.Enclave,
 	error,
 ) {
-	matchingEnclavesByNamespace, err := backend.getMatchingEnclaves(ctx, filters)
+	matchingEnclaves, _, err := backend.getMatchingEnclaveObjectsAndKubernetesResources(ctx, filters)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting enclaves matching the following filters: %+v", filters)
 	}
-
-	matchingEnclavesByEnclaveId := map[enclave.EnclaveID]*enclave.Enclave{}
-	for _, enclaveObj := range matchingEnclavesByNamespace {
-		matchingEnclavesByEnclaveId[enclaveObj.GetID()] = enclaveObj
-	}
-
-	return matchingEnclavesByEnclaveId, nil
+	return matchingEnclaves, nil
 }
 
-func (backend *KubernetesKurtosisBackend) StopEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (successfulEnclaveIds map[enclave.EnclaveID]bool, erroredEnclaveIds map[enclave.EnclaveID]error, resultErr error) {
-	//TODO implement me
-	panic("implement me")
+func (backend *KubernetesKurtosisBackend) StopEnclaves(
+	ctx context.Context,
+	filters *enclave.EnclaveFilters,
+) (
+	map[enclave.EnclaveID]bool,
+	map[enclave.EnclaveID]error,
+	error,
+) {
+
+	_, matchingKubernetesResources, err := backend.getMatchingEnclaveObjectsAndKubernetesResources(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting enclaves and Kubernetes resources matching filters '%+v'", filters)
+	}
+
+	successfulEnclaveIds := map[enclave.EnclaveID]bool{}
+	erroredEnclaveIds := map[enclave.EnclaveID]error{}
+	for enclaveId, resources := range matchingKubernetesResources {
+		enclaveIdStr := string(enclaveId)
+		if resources.namespace != nil {
+			namespaceName := resources.namespace.GetName()
+
+			enclaveWithIDMatchLabels := map[string]string{
+				label_key_consts.AppIDLabelKey.GetString():     label_value_consts.AppIDLabelValue.GetString(),
+				label_key_consts.EnclaveIDLabelKey.GetString(): enclaveIdStr,
+			}
+
+			// Services
+			servicesByEnclaveId, err := kubernetes_resource_collectors.CollectMatchingServices(
+				ctx,
+				backend.kubernetesManager,
+				namespaceName,
+				enclaveWithIDMatchLabels,
+				label_key_consts.EnclaveIDLabelKey.GetString(),
+				map[string]bool{
+					enclaveIdStr: true,
+				},
+			)
+			if err != nil {
+				return nil, nil, stacktrace.Propagate(err, "An error occurred getting services matching enclave ID '%v' in namespace '%v'", enclaveIdStr, namespaceName)
+			}
+			services, found := servicesByEnclaveId[enclaveIdStr]
+			if !found {
+				erroredEnclaveIds[enclaveId] = stacktrace.NewError("Expected to find enclave's services for enclave '%v' in services by enclave ID map '%+v' but was not found, this is a bug in Kurtosis", enclaveIdStr, servicesByEnclaveId)
+			}
+
+			if services != nil {
+				errorsByServiceName := map[string]error{}
+				for _, service := range services {
+					serviceName := service.GetName()
+					if err := backend.kubernetesManager.RemoveSelectorsFromService(ctx, namespaceName, serviceName); err != nil {
+						errorsByServiceName[serviceName] = err
+						continue
+					}
+				}
+
+				if len(errorsByServiceName) > 0 {
+					combinedErrorTitle := fmt.Sprintf("Namespace %v - Service", namespaceName)
+					combinedError := buildCombinedError(errorsByServiceName, combinedErrorTitle)
+					erroredEnclaveIds[enclaveId] = stacktrace.Propagate(
+						combinedError,
+						"An error occurred removing one or more service's selectors in namespace '%v' for enclave with ID '%v'",
+						namespaceName,
+						enclaveId,
+					)
+					continue
+				}
+			}
+
+			// Pods
+			if resources.pods != nil {
+				errorsByPodName := map[string]error{}
+				for _, pod := range resources.pods {
+					podName := pod.GetName()
+					if err := backend.kubernetesManager.RemovePod(ctx, namespaceName, podName); err != nil {
+						errorsByPodName[podName] = err
+						continue
+					}
+				}
+
+				if len(errorsByPodName) > 0 {
+					combinedErrorTitle := fmt.Sprintf("Namespace %v - Pod", namespaceName)
+					combinedError := buildCombinedError(errorsByPodName, combinedErrorTitle)
+					erroredEnclaveIds[enclaveId] = stacktrace.Propagate(
+						combinedError,
+						"An error occurred removing one or more pods in namespace '%v' for enclave with ID '%v'",
+						namespaceName,
+						enclaveId,
+					)
+					continue
+				}
+			}
+
+			successfulEnclaveIds[enclaveId] = true
+		}
+	}
+
+	return successfulEnclaveIds, erroredEnclaveIds, nil
 }
 
 func (backend *KubernetesKurtosisBackend) DumpEnclave(ctx context.Context, enclaveId enclave.EnclaveID, outputDirpath string) error {
-	namespace, err := backend.getEnclaveNamespace(ctx, enclaveId)
+	_, kubernetesResources, err := backend.getSingleEnclaveAndKubernetesResources(ctx, enclaveId)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting enclave namespace for enclave ID '%v'", enclaveId)
+		return stacktrace.Propagate(err, "An error occurred getting enclave object and Kubernetes resources for enclave ID '%v'", enclaveId)
 	}
-	namespaceName := namespace.Name
+	namespace := kubernetesResources.namespace
+	if namespace == nil {
+		return stacktrace.NewError("Cannot dump enclave '%v' because no Kubernetes namespace exists for it", enclaveId)
+	}
 
 	// Create output directory
 	if _, err := os.Stat(outputDirpath); !os.IsNotExist(err) {
@@ -191,14 +303,14 @@ func (backend *KubernetesKurtosisBackend) DumpEnclave(ctx context.Context, encla
 		return stacktrace.Propagate(err, "An error occurred creating output directory at '%v'", outputDirpath)
 	}
 
-	enclavePods, err := backend.getAllEnclavePods(ctx, namespaceName, enclaveId)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting pods in enclave '%v'", enclaveId)
+	podsToDump := kubernetesResources.pods
+	if podsToDump == nil {
+		podsToDump = []*apiv1.Pod{}
 	}
 
 	workerPool := workerpool.New(numPodsToDumpAtOnce)
-	resultErrsChan := make(chan error, len(enclavePods))
-	for _, pod := range enclavePods {
+	resultErrsChan := make(chan error, len(podsToDump))
+	for _, pod := range podsToDump {
 		/*
 			!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 			It's VERY important that the actual `func()` job function get created inside a helper function!!
@@ -215,7 +327,7 @@ func (backend *KubernetesKurtosisBackend) DumpEnclave(ctx context.Context, encla
 		jobToSubmit := createDumpPodJob(
 			ctx,
 			backend.kubernetesManager,
-			namespaceName,
+			namespace.Name,
 			pod,
 			outputDirpath,
 			resultErrsChan,
@@ -255,83 +367,196 @@ func (backend *KubernetesKurtosisBackend) DestroyEnclaves(ctx context.Context, f
 // ====================================================================================================
 //                                     Private Helper Methods
 // ====================================================================================================
-// Gets enclaves matching the search filters, indexed by their [namespace]
-func (backend *KubernetesKurtosisBackend) getMatchingEnclaves(
+func (backend *KubernetesKurtosisBackend) getMatchingEnclaveObjectsAndKubernetesResources(
 	ctx context.Context,
 	filters *enclave.EnclaveFilters,
 ) (
-	map[string]*enclave.Enclave,
+	map[enclave.EnclaveID]*enclave.Enclave,
+	map[enclave.EnclaveID]*enclaveKubernetesResources,
 	error,
 ) {
-	matchingEnclaves := map[string]*enclave.Enclave{}
-
-	enclaveNamespaces, err := backend.getAllEnclaveNamespaces(ctx)
+	matchingResources, err := backend.getMatchingEnclaveKubernetesResources(ctx, filters.IDs)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting all enclave namespaces")
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting enclave Kubernetes resources matching IDs: %+v", filters.IDs)
 	}
 
-	for _, enclaveNamespace := range enclaveNamespaces {
-		enclaveNamespaceName := enclaveNamespace.GetName()
-		enclaveNamespaceLabels := enclaveNamespace.GetLabels()
+	enclaveObjects, err := getEnclaveObjectsFromKubernetesResources(matchingResources)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting enclave objects from Kubernetes resources")
+	}
 
-		enclaveIdStr, found := enclaveNamespaceLabels[label_key_consts.EnclaveIDLabelKey.GetString()]
-		if !found {
-			return nil, stacktrace.NewError("Expected to find a label with name '%v' in Kubernetes namespace '%v', but no such label was found", label_key_consts.EnclaveIDLabelKey.GetString(), enclaveNamespaceName)
-		}
-		enclaveId := enclave.EnclaveID(enclaveIdStr)
-		// If the IDs filter is specified, drop enclaves not matching it
+	// Finally, apply the filters
+	resultEnclaveObjs := map[enclave.EnclaveID]*enclave.Enclave{}
+	resultKubernetesResources := map[enclave.EnclaveID]*enclaveKubernetesResources{}
+	for enclaveId, enclaveObj := range enclaveObjects {
 		if filters.IDs != nil && len(filters.IDs) > 0 {
-			if _, found := filters.IDs[enclaveId]; !found {
+			if _, found := filters.IDs[enclaveObj.GetID()]; !found {
 				continue
 			}
 		}
 
-		enclavePods, err := backend.getAllEnclavePods(ctx, enclaveNamespaceName, enclaveId)
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred getting all enclave pods for enclave '%v' in namespace '%v'", enclaveId, enclaveNamespaceName)
-		}
-
-		enclaveStatus, err := getEnclaveStatusFromEnclavePods(enclavePods)
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred getting enclave status from enclave pods '%+v'", enclavePods)
-		}
-
-		// If the Statuses filter is specified, drop enclaves not matching it
 		if filters.Statuses != nil && len(filters.Statuses) > 0 {
-			if _, found := filters.Statuses[enclaveStatus]; !found {
+			if _, found := filters.Statuses[enclaveObj.GetStatus()]; !found {
 				continue
 			}
 		}
 
-		enclaveObj := enclave.NewEnclave(enclaveId, enclaveStatus)
-
-		matchingEnclaves[enclaveNamespaceName] = enclaveObj
+		resultEnclaveObjs[enclaveId] = enclaveObj
+		if _, found := matchingResources[enclaveId]; !found {
+			return nil, nil, stacktrace.NewError("Expected to find Kubernetes resources matching enclave '%v' but none was found", enclaveId)
+		}
+		// Okay to do because we're guaranteed a 1:1 mapping between enclave_obj:enclave_resources
+		resultKubernetesResources[enclaveId] = matchingResources[enclaveId]
 	}
 
-	return matchingEnclaves, nil
+	return resultEnclaveObjs, resultKubernetesResources, nil
 }
 
-func (backend *KubernetesKurtosisBackend) getAllEnclavePods(ctx context.Context, enclaveNamespaceName string, enclaveId enclave.EnclaveID) ([]apiv1.Pod, error) {
-	matchingPods := []apiv1.Pod{}
-
-	matchLabels := map[string]string{
-		label_key_consts.AppIDLabelKey.GetString():     label_value_consts.AppIDLabelValue.GetString(),
-		label_key_consts.EnclaveIDLabelKey.GetString(): string(enclaveId),
+func (backend *KubernetesKurtosisBackend) getSingleEnclaveAndKubernetesResources(ctx context.Context, enclaveId enclave.EnclaveID) (*enclave.Enclave, *enclaveKubernetesResources, error) {
+	enclaveSearchFilters := &enclave.EnclaveFilters{
+		IDs: map[enclave.EnclaveID]bool{
+			enclaveId: true,
+		},
 	}
-
-	foundPods, err := backend.kubernetesManager.GetPodsByLabels(ctx, enclaveNamespaceName, matchLabels)
+	matchingEnclaveObjects, matchingKubernetesResources, err := backend.getMatchingEnclaveObjectsAndKubernetesResources(ctx, enclaveSearchFilters)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting pods by labels '%+v' in namespace '%v'", matchLabels, enclaveNamespaceName)
+		return nil, nil, stacktrace.Propagate(err, "An error occurred getting enclave objects and Kubernetes resources matching enclave '%v'", enclaveId)
+	}
+	if len(matchingEnclaveObjects) == 0 || len(matchingKubernetesResources) == 0 {
+		return nil, nil, stacktrace.NewError("Didn't find enclave objects and Kubernetes resources for enclave '%v'", enclaveId)
+	}
+	if len(matchingEnclaveObjects) > 1 || len(matchingEnclaveObjects) > 0 {
+		return nil, nil, stacktrace.NewError("Found more than one enclave objects/Kubernetes resources for enclave '%v'", enclaveId)
 	}
 
-	if foundPods.Items != nil {
-		matchingPods = foundPods.Items
+	enclaveObject, found := matchingEnclaveObjects[enclaveId]
+	if !found {
+		return nil, nil, stacktrace.NewError("No enclave object exists for enclave '%v'", enclaveId)
 	}
-	return matchingPods, nil
+
+	kubernetesResources, found := matchingKubernetesResources[enclaveId]
+	if !found {
+		return nil, nil, stacktrace.NewError("No Kubernetes resources object exists for enclave '%v'", enclaveId)
+	}
+
+	return enclaveObject, kubernetesResources, nil
 }
 
-func getEnclaveStatusFromEnclavePods(enclavePods []apiv1.Pod) (enclave.EnclaveStatus, error) {
-	resultEnclaveStatus := enclave.EnclaveStatus_Stopped
+// Get back any and all enclave's Kubernetes resources matching the given enclave IDs, where a nil or empty map == "match all enclave IDs"
+func (backend *KubernetesKurtosisBackend) getMatchingEnclaveKubernetesResources(ctx context.Context, enclaveIds map[enclave.EnclaveID]bool) (
+	map[enclave.EnclaveID]*enclaveKubernetesResources,
+	error,
+) {
+
+	result := map[enclave.EnclaveID]*enclaveKubernetesResources{}
+
+	enclaveMatchLabels := getEnclaveMatchLabels()
+
+	enclaveIdsStrSet := map[string]bool{}
+	for enclaveId, booleanValue := range enclaveIds {
+		enclaveIdStr := string(enclaveId)
+		enclaveIdsStrSet[enclaveIdStr] = booleanValue
+	}
+
+	// Namespaces
+	namespaces, err := kubernetes_resource_collectors.CollectMatchingNamespaces(
+		ctx,
+		backend.kubernetesManager,
+		enclaveMatchLabels,
+		label_key_consts.EnclaveIDLabelKey.GetString(),
+		enclaveIdsStrSet,
+	)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting enclave namespaces matching IDs '%+v'", enclaveIdsStrSet)
+	}
+
+	// Per-namespace objects
+	for enclaveIdStr, namespacesForEnclaveId := range namespaces {
+		if len(namespacesForEnclaveId) == 0 {
+			return nil, stacktrace.NewError(
+				"Ostensibly found namespaces for enclave ID '%v', but no namespace objects were returned",
+				enclaveIdStr,
+			)
+		}
+		if len(namespacesForEnclaveId) > 1 {
+			return nil, stacktrace.NewError(
+				"Expected at most one namespace to match enclave ID '%v', but got '%v'",
+				enclaveIdStr,
+				len(namespacesForEnclaveId),
+			)
+		}
+
+		namespace := namespacesForEnclaveId[0]
+
+		enclaveWithIDMatchLabels := map[string]string{
+			label_key_consts.AppIDLabelKey.GetString():     label_value_consts.AppIDLabelValue.GetString(),
+			label_key_consts.EnclaveIDLabelKey.GetString(): enclaveIdStr,
+		}
+
+		// Pods
+		pods, err := kubernetes_resource_collectors.CollectMatchingPods(
+			ctx,
+			backend.kubernetesManager,
+			namespace.GetName(),
+			enclaveWithIDMatchLabels,
+			label_key_consts.EnclaveIDLabelKey.GetString(),
+			map[string]bool{
+				enclaveIdStr: true,
+			},
+		)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting pods matching enclave ID '%v' in namespace '%v'", enclaveIdStr, namespace.GetName())
+		}
+
+		enclaveId := enclave.EnclaveID(enclaveIdStr)
+
+		enclaveResources, found := result[enclaveId]
+		if !found {
+			enclaveResources = &enclaveKubernetesResources{}
+		}
+		enclaveResources.namespace = namespace
+		enclaveResources.pods = pods[enclaveIdStr]
+
+		result[enclaveId] = enclaveResources
+	}
+
+	return result, nil
+}
+
+func getEnclaveObjectsFromKubernetesResources(
+	allResources map[enclave.EnclaveID]*enclaveKubernetesResources,
+) (
+	map[enclave.EnclaveID]*enclave.Enclave,
+	error,
+) {
+	result := map[enclave.EnclaveID]*enclave.Enclave{}
+
+	for enclaveId, resourcesForEnclaveId := range allResources {
+
+		if resourcesForEnclaveId.namespace == nil {
+			return nil, stacktrace.NewError("Cannot create an enclave object '%v' when no Kubernetes namespace exists", enclaveId)
+		}
+
+		enclaveStatus, err := getEnclaveStatusFromEnclavePods(resourcesForEnclaveId.pods)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting enclave status from enclave pods '%+v'", resourcesForEnclaveId.pods)
+		}
+
+		enclaveObj := enclave.NewEnclave(
+			enclaveId,
+			enclaveStatus,
+		)
+
+		result[enclaveId] = enclaveObj
+	}
+	return result, nil
+}
+
+func getEnclaveStatusFromEnclavePods(enclavePods []*apiv1.Pod) (enclave.EnclaveStatus, error) {
+	resultEnclaveStatus := enclave.EnclaveStatus_Empty
+	if len(enclavePods) > 0 {
+		resultEnclaveStatus = enclave.EnclaveStatus_Stopped
+	}
 	for _, enclavePod := range enclavePods {
 		podPhase := enclavePod.Status.Phase
 
@@ -346,6 +571,7 @@ func getEnclaveStatusFromEnclavePods(enclavePods []apiv1.Pod) (enclave.EnclaveSt
 			break
 		}
 	}
+
 	return resultEnclaveStatus, nil
 }
 
@@ -353,7 +579,7 @@ func createDumpPodJob(
 	ctx context.Context,
 	kubernetesManager *kubernetes_manager.KubernetesManager,
 	namespaceName string,
-	pod apiv1.Pod,
+	pod *apiv1.Pod,
 	enclaveOutputDirpath string,
 	resultErrsChan chan error,
 ) func() {
@@ -372,7 +598,7 @@ func dumpPodInfo(
 	ctx context.Context,
 	kubernetesManager *kubernetes_manager.KubernetesManager,
 	namespaceName string,
-	pod apiv1.Pod,
+	pod *apiv1.Pod,
 	enclaveOutputDirpath string,
 ) error {
 	podName := pod.Name