+++ container-engine-lib/docs/changelog.md
@@ -1,13 +1,12 @@
 # TBD
 ### Features
-* Build out `KubernetesKurtosisBackend.DumpEnclave`
-
-### Features
+* Added `KubernetesKurtosisBackend.StopEnclaves` functionality to kubernetes backend
 * Added `KubernetesManager.IsPersistentVolumeClaimBound` to check when a Persistent Volume has been bound to a Persistent Volume Claim
 * Updated `KubernetesManager.CreatePersistentVolumeClaim` to wait for the PersistentVolumeClaim to get bound
 * Added `CollectMatchingRoles` and `CollectMatchingRoleBindings` in `kubernetes_resource_collectors` package
 * Upped the CircleCI resource class to 'large' since builds are 1m30s and CircleCI showed that we're maxing out the CPU
 * Added a build cache to each build
+* Build out `KubernetesKurtosisBackend.DumpEnclave`
 
 ### Fixes
 * Added apiv1 prefix to `KubernetesManager.GetPodPortforwardEndpointUrl`